feat: refactor database imports to use @hyperlocal/db and remove unused services
