+++ apps/admin/app/routes/_auth+/auth.google.callback.ts
@@ -17,7 +17,5 @@ export const loader = async ({ request, context }: LoaderFunctionArgs) => {
     'Set-Cookie': await sessionStorage.commitSession(session),
   })
 
-  console.log(Object.fromEntries(headers.entries()))
-
   return redirect('/', { headers })
 }
