+++ apps/admin/package.json
@@ -71,11 +71,9 @@
     "@biomejs/biome": "catalog:",
     "@cloudflare/workers-types": "4.20241106.0",
     "@react-router/dev": "catalog:",
-    "@types/debug": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
     "autoprefixer": "catalog:",
-    "debug": "catalog:",
     "npm-run-all": "4.1.5",
     "postcss": "catalog:",
     "prettier": "catalog:",
