+++ packages/db/src/index.ts
@@ -1,36 +1,36 @@
 import { LibsqlDialect } from '@libsql/kysely-libsql'
-import createDebug from 'debug'
-import {
-  CamelCasePlugin,
-  DeduplicateJoinsPlugin,
-  Kysely,
-  ParseJSONResultsPlugin,
-} from 'kysely'
+import { CamelCasePlugin, Kysely, ParseJSONResultsPlugin } from 'kysely'
+import { getSessionContext } from 'session-context'
 import type { DB } from './schema'
 export { sql } from 'kysely'
 export type { Insertable, Selectable, Updateable } from 'kysely'
 export type * from './types'
 export type { DB }
 
-const debug = createDebug('app:db')
-
-const getDatabaseConfig = () => {
-  const url = process.env.DATABASE_URL
-  const authToken = process.env.TURSO_AUTH_TOKEN
-  if (!url || !authToken) {
-    throw new Error(
-      'Missing required database configuration. Please check DATABASE_URL and TURSO_AUTH_TOKEN environment variables.',
-    )
+export const getDb = () => {
+  const store = getSessionContext<{
+    db?: Kysely<DB>
+    env: { DATABASE_URL: string; TURSO_AUTH_TOKEN: string }
+  }>()
+  if (!store.db) {
+    store.db = new Kysely<DB>({
+      dialect: new LibsqlDialect({
+        url: store.env.DATABASE_URL ?? '',
+        authToken: store.env.TURSO_AUTH_TOKEN ?? '',
+      }),
+      plugins: [new CamelCasePlugin(), new ParseJSONResultsPlugin()],
+    })
   }
-  return { url, authToken }
+  return store.db
 }
 
-export const db = new Kysely<DB>({
-  dialect: new LibsqlDialect(getDatabaseConfig()),
-  log: (event) => debug([event.query.sql, event.query.parameters]),
-  plugins: [
-    new CamelCasePlugin(),
-    new ParseJSONResultsPlugin(),
-    new DeduplicateJoinsPlugin(),
-  ],
+export const db = new Proxy<Kysely<DB>>({} as never, {
+  get(_target: unknown, props: keyof Kysely<DB>) {
+    const instance = getDb()
+    const value = instance[props]
+    if (typeof value === 'function') {
+      return value.bind(instance)
+    }
+    return value
+  },
 })
