+++ pnpm-workspace.yaml
@@ -27,7 +27,6 @@ catalog:
   '@react-router/cloudflare': 7.0.0-pre.5
   '@react-router/remix-routes-option-adapter': 7.0.0-pre.5
   '@react-router/serve': 7.0.0-pre.5
-  '@types/debug': 4.1.12
   '@types/node': 22.9.0
   '@types/nprogress': 0.2.3
   '@types/react': 18.3.12
@@ -42,7 +41,6 @@ catalog:
   concurrently: 9.1.0
   crawlee: 3.12.0
   date-fns: 4.1.0
-  debug: 4.3.7
   dotenv: 16.4.5
   duckdb: 1.1.3
   embla-carousel-react: 8.3.1
